pyproject.toml
[project]
name = "goal-apps"
version="0.0.1"
description = "High-level applications built on GOAL"
dependencies = [
    "goal-jax",
    "matplotlib",
    "scikit-learn",
    "seaborn",
    "hydra-core",
    "wandb",
    "typer",
    "pandas",
    "h5py",
    "tables",
]

[project.scripts]
goal = "apps.cli.main:main"

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pytest-cov>=4.0",  # For coverage reporting
]

gpu = [
    "jax[cuda12]",
]

datasets = [
    "torchvision",
]

[build-system]
# These are standard for modern Python projects:
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
requires-python = ">=3.12"
# Add your dependencies:
[tool.setuptools.packages.find]
where = ["src", "."]
include =["apps*", "plugins*"]

[tool.setuptools.package-data]
goal = ["py.typed"]

[tool.pyright]
include = ["src/apps", "plugins"]
typeCheckingMode = "recommended"
venvPath = "."
venv = ".venv"
reportUnknownParameterType = "none"
reportUnknownMemberType = "none"
reportUnknownArgumentType = "none"
reportAny = "none"
reportExplicitAny = "none"
reportMissingSuperCall = "none"
reportUnusedCallResult = "none" 

[tool.ruff]
src = ["src"]
output-format="concise"
exclude = ["node_modules", "__pycache__"]
lint.select = [
    "F",  # Pyflakes
    "E",  # PEP8
    "W",  # PEP8
    "C90",  # McCabe
    "D",  # Pydocstyle
    "R",  # Ruff-specific
    "N",  # Naming
    "I",  # Import conventions
    "SIM",  # flake8-simplify
    "UP",  # pyupgrade
    "RUF",  # ruff-specific rules
    "ARG",  # flake8-unused-arguments
    "TRY",  # tryceratops
    "PLR0402", # import a.b as b -> from a import b
]

lint.ignore = [
    "E501",  # Example: Ignore line length warnings
    "D",     # Ignore all docstring-related warnings
    "ARG002",# Ignore unused-argument in method > problematic eg when required by parent class / interface
    "ARG003",# Unused parameters
    "TRY003",# Allow to write exception messages in place instead of subclassing
    # "RUF001",# Allow confusable characters (eg σ)
    "F722",  # spurious errors because of jaxtyping
]

[tool.ruff.format]
docstring-code-format = true
